devik go implementacion y set up inicial

+++ src/endpoints/db/clientes/useDeleteClient.ts
@@ -1,5 +1,5 @@
 import {db} from '@/firebase';
-import {deleteDoc, doc, getDoc} from 'firebase/firestore';
+import {collection, deleteDoc, doc, getDoc, getDocs} from 'firebase/firestore';
 import {useCallback, useState} from 'react';
 import {FIRESTORE_CLIENTES_PATH} from '@/constants';
 import {Cliente} from '@/types';
@@ -11,6 +11,13 @@ export default function useDeleteClient() {
   const [isLoadingDeleteCliente, setIsLoadingDeleteCliente] = useState<boolean>(false);
   const {deleteFile} = useDeleteFile();
 
+  const deleteSubcollectionDocs = useCallback(async (clienteId: string, subcollectionName: string) => {
+    const subcollectionRef = collection(db, FIRESTORE_CLIENTES_PATH, clienteId, subcollectionName);
+    const snapshot = await getDocs(subcollectionRef);
+    const deletePromises = snapshot.docs.map((doc) => deleteDoc(doc.ref));
+    await Promise.all(deletePromises);
+  }, []);
+
   const deleteCliente = useCallback(
     async (clienteId: string): Promise<void> => {
       setIsLoadingDeleteCliente(true);
@@ -18,9 +25,19 @@ export default function useDeleteClient() {
       try {
         const clienteDoc = await getDoc(clienteRef);
         const clienteData = clienteDoc.data() as Cliente;
-        await deleteDoc(clienteRef);
+        if (!clienteDoc.exists()) {
+          toast.error('El cliente no existe o ya fue eliminado.');
+          return;
+        }
+        const subCollections = (clienteData.tiposRegistros ?? []).map(({tipo}) => tipo.toUpperCase());
+        const deletePromises = subCollections.map((subcollection) => deleteSubcollectionDocs(clienteId, subcollection));
+        await Promise.all(deletePromises);
+
         if (clienteData.logo) await deleteFile(clienteData.logo);
         if (clienteData.documento) await deleteFile(clienteData.documento);
+
+        await deleteDoc(clienteRef);
+
         toast.success(`Se ha eliminado el cliente ${clienteData.nombre} correctamente`);
       } catch (error: any) {
         toast.error('¡Ups ha ocurrido un error, intenta de nuevo más tarde!');
@@ -29,8 +46,9 @@ export default function useDeleteClient() {
         setIsLoadingDeleteCliente(false);
       }
     },
-    [deleteFile]
+    [deleteFile, deleteSubcollectionDocs]
   );
+
   return {
     isLoadingDeleteCliente,
     deleteCliente
